Rename pretty_errors->errors
