+++ pkg/errors/error_test.go
@@ -1,4 +1,4 @@
-package pretty_errors
+package errors
 
 import (
 	"errors"
