+++ pkg/errors/text_formatter_test.go
@@ -1,4 +1,4 @@
-package pretty_errors
+package errors
 
 import "testing"
